productratechanges

+++ MangalWeb.Model/Masters/ProductRateViewModel.cs
@@ -16,7 +16,7 @@ namespace MangalWeb.Model.Masters
         public string ProductRateDate { get; set; }
 
         [Required(ErrorMessage = "Please Select Product")]
-        public short Product { get; set; }
+        public int Product { get; set; }
         public string ProductStr { get; set; }
 
         public string operation { get; set; }
